+++ python/fastapi/src/alpha93/fastapi/_contextlib.py
@@ -1,2 +1,0 @@
-from contextlib import AsyncExitStack
-__all__ = ("AsyncExitStack",)