+++ python/fastapi/src/alpha93/fastapi/_internal/_compat/shared.py
@@ -1,2 +1,0 @@
-__all__ = ("lenient_issubclass",)
-from pydantic._internal._utils import lenient_issubclass