+++ python/fastapi/src/alpha93/fastapi/_internal/routing/_route/__init__.py
@@ -1,6 +1,0 @@
-__all__ = ("APIRoute", "APIWebSocketRoute", "Mount",)
-
-from starlette.routing import Mount
-
-from ._route import APIRoute
-from ._websocket import APIWebSocketRoute