+++ python/hatch-shadow/src/hatch_shadow/hooks.py
@@ -1,8 +1,0 @@
-from hatchling.plugin import hookimpl
-
-from .builder import ShadowBuildHook
-
-
-@hookimpl
-def hatch_register_build_hook():
-    return ShadowBuildHook