+++ python/hatch-shadow/src/hatch_shadow/managers/__init__.py
@@ -1,4 +1,0 @@
-from ._managers import PackageManager
-
-
-__all__ = ("PackageManager",)