+++ python/makepatch/src/makepatch/__hooks__.py
@@ -1,7 +1,0 @@
-from hatchling.plugin import hookimpl
-from makepatch.hooks import PatcherBuildHook
-
-
-@hookimpl
-def hatch_register_build_hook():
-    return PatcherBuildHook